migrate: split into migrate and migrate_container
