+++ lxd/container_snapshot.go
@@ -246,7 +246,7 @@ func snapshotPost(d *Daemon, r *http.Request, sc container, containerName string
 
 		if req.Target != nil {
 			// Push mode
-			err := ws.ConnectTarget(*req.Target)
+			err := ws.ConnectContainerTarget(*req.Target)
 			if err != nil {
 				return InternalError(err)
 			}
